refactored revision page and wiki page to have single source of truth
